python/requirements.txt
jupyterlab
jupyterlab_code_formatter
autopep8
tensorflow
sklearn
matplotlib
requests
dlib
pillow
